Changing from add/minus actions to saving states
In every well object, I’ve had the structure been to save all solutions
and actions in one actionList. I could then add or subtract all of
these solutions to figure out the state of the well at a certain point
in time. This addition and subtraction and keeping track of what’s been
added or subtracted is just too much work. Instead, I plan on adding a
$states slot to well objects to keep track of the added actions that
describe the state of the well at specific points in time. The $actions
slot will still hold the list of actions that were actually performed.
This is primarily sacrificing memory for simplicity.